Refactor exploiters into plugins. Change configuration to support it and remove unneeded complex importing. Changed main monkey code to support exploiter plugins.
