+++ monkey/infection_monkey/exploit/win_ms08_067.py
@@ -19,7 +19,7 @@ from infection_monkey.exploit.tools.smb_tools import SmbTools
 from infection_monkey.model import DROPPER_CMDLINE_WINDOWS, MONKEY_CMDLINE_WINDOWS
 from infection_monkey.network.smbfinger import SMBFinger
 from infection_monkey.network.tools import check_tcp_port
-from . import HostExploiter
+from infection_monkey.exploit.HostExploiter import HostExploiter
 
 LOG = getLogger(__name__)
